+++ index.html
@@ -8,6 +8,8 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
+  <script type="module" src="js/main.js"></script>
+  <script src="js/functions.js"></script>
 </head>
 
 <body>
@@ -242,8 +244,8 @@
     </section>
   </template>
 
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+
+
 </body>
 
 </html>
